Move common handler tests logic to common http test file
