+++ go-backend/service/common_http_test.go
@@ -1,0 +1,35 @@
+package service
+
+import (
+	"testing"
+	"github.com/stretchr/testify/suite"
+	"net/http"
+	"net/http/httptest"
+	"strings"
+	"github.com/gorilla/mux"
+)
+
+func TestExampleTestSuite(t *testing.T) {
+	suite.Run(t, new(UsersHandlerTestSuite))
+	suite.Run(t, new(OrganizationHandlerTestSuite))
+	suite.Run(t, new(RecognitionHi5HandlerTestSuite))
+	suite.Run(t, new(CoreValueHandlerTestSuite))
+}
+
+// path: is used to configure router path (eg: /users/{id})
+// requestURL: current request path (eg: /users/1)
+func makeHTTPCall(method, path, requestURL, body string, handlerFunc http.HandlerFunc) (recorder *httptest.ResponseRecorder) {
+	// create a http request using the given parameters
+	req, _ := http.NewRequest(method, requestURL, strings.NewReader(body))
+
+	// test recorder created for capturing api responses
+	recorder = httptest.NewRecorder()
+
+	// create a router to serve the handler in test with the prepared request
+	router := mux.NewRouter()
+	router.HandleFunc(path, handlerFunc).Methods(method)
+
+	// serve the request and write the response to recorder
+	router.ServeHTTP(recorder, req)
+	return
+}
